+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 Here you can see the full list of changes between each SQLAlchemy-Utils release.
 
 
+0.30.9 (2015-06-05)
+^^^^^^^^^^^^^^^^^^^
+
+- Added get_type utility function
+
+
 0.30.8 (2015-06-05)
 ^^^^^^^^^^^^^^^^^^^